Gets real decimal value for UTF32 characters (e.g. emoji)
Changes the method of getting the decimal of a character from turning it directly into an integer to getting its code point at index 0, which means the code point of the first character of an input string (if the input is "34", it will return all values for character "3"). This method works for all character since the resulting integer is a unicode code point instead of a UTF16 character integer representation.